requirements.txt
# ===== CORE FRAMEWORKS =====
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# ===== AI & ML =====
openai==1.59.2
torch==2.10.0
transformers==5.3.0

# ===== VOICE =====
elevenlabs==2.0.0
gtts==2.5.4
pydub==0.25.1

# ===== FRONTEND =====
streamlit==1.41.1
plotly==5.24.1
pandas==2.3.3
numpy==2.4.3

# ===== UTILITIES =====
python-multipart==0.0.12
requests==2.32.3
setuptools==75.6.0
wheel==0.45.1
schedule==1.2.2
python-dateutil==2.9.0
click==8.1.8
apscheduler==3.11.2
httpx==0.28.1